+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Remove probably superfluous check for unsupported $ref inside parameter schema
+
 ## [0.2.1] - 2023-03-31
 
 - Fix links in gemspec